use glam Vec3

+++ src/ray.rs
@@ -1,4 +1,5 @@
-use crate::vec3::{Point3, Vec3};
+use crate::vec3::Point3;
+use glam::Vec3;
 
 pub struct Ray {
     pub origin: Point3,
@@ -10,7 +11,7 @@ impl Ray {
         Self { origin, direction }
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: f32) -> Point3 {
         self.origin + t * self.direction
     }
 }
